Added authorizaiton to all the post APIs

+++ src/main/java/com/bhargav/blog/blog_api/controller/PostController.java
@@ -1,7 +1,9 @@
 package com.bhargav.blog.blog_api.controller;
 
+import com.bhargav.blog.blog_api.config.JwtUtil;
 import com.bhargav.blog.blog_api.model.Post;
 import com.bhargav.blog.blog_api.service.PostService;
+import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -20,19 +22,35 @@ public class PostController {
 
     private final PostService postService;
 
+    private final JwtUtil jwtUtil;
+
     @PostMapping("/create")
-    public ResponseEntity<String> createPost(@RequestBody Post post) {
+    public ResponseEntity<String> createPost(@RequestBody Post post, HttpServletRequest request) {
+
+        if(!jwtUtil.isAuthorized(request)){
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         postService.createPost(post);
         return new ResponseEntity<>("Post created succesfully!!!", HttpStatus.CREATED);
     }
 
     @PutMapping("/delete/{id}")
-    public ResponseEntity<String> deletePost(@PathVariable int id) {
+    public ResponseEntity<String> deletePost(@PathVariable int id, HttpServletRequest request) {
+
+        if(!jwtUtil.isAuthorized(request)){
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
+
         return postService.deletePost(id);
     }
 
     @PutMapping("/deleteAll")
-    public ResponseEntity<String> deleteAllPosts() {
+    public ResponseEntity<String> deleteAllPosts(HttpServletRequest request) {
+
+        if(!jwtUtil.isAuthorized(request)){
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
+
         postService.deleteAllPosts();
         return new ResponseEntity<>("All posts deleted succesfully!", HttpStatus.OK);
     }
@@ -40,7 +58,12 @@ public class PostController {
 
     @GetMapping("/get")
     public ResponseEntity<Object> getAllPosts(@RequestParam(defaultValue = "0") int page,
-                                                  @RequestParam(defaultValue = "5") int limit) {
+                                                  @RequestParam(defaultValue = "5") int limit,
+                                              HttpServletRequest request) {
+
+        if(!jwtUtil.isAuthorized(request)){
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
 
         Pageable pageable = PageRequest.of(page, limit);
 
@@ -58,7 +81,13 @@ public class PostController {
     }
 
     @GetMapping("/search")
-    public ResponseEntity<Object> searchPosts(@RequestParam String tag) {
+    public ResponseEntity<Object> searchPosts(@RequestParam String tag, HttpServletRequest request) {
+
+        if(!jwtUtil.isAuthorized(request)){
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
+
+
         List<Post> posts =  postService.searchPosts(tag);
 
         if (posts.isEmpty()) {
@@ -73,7 +102,12 @@ public class PostController {
     }
 
     @PutMapping("/update/{id}")
-    public ResponseEntity<String> updatePost(@PathVariable int id, @RequestBody Post post) {
+    public ResponseEntity<String> updatePost(@PathVariable int id, @RequestBody Post post, HttpServletRequest request) {
+
+        if(!jwtUtil.isAuthorized(request)){
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
+
         boolean updated = postService.updatePost(id, post);
 
         if (updated) {
